refactor: コンポーネント化した->AIにすごくやってもらった
