+++ PrefeFortune/Components/BloodTypePickerView.swift
@@ -1,0 +1,26 @@
+//
+//  BloodTypePickerView.swift
+//  PrefeFortune
+//
+//  Created by 櫻井絵理香 on 2024/10/05.
+//
+
+import SwiftUI
+
+struct BloodTypePickerView: View {
+    @Binding var bloodType: String
+    let bloodTypes: [String]
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 5) {
+            Text("血液型")
+                .font(.headline)
+            Picker("血液型を選択してください", selection: $bloodType) {
+                ForEach(bloodTypes, id: \.self) { type in
+                    Text(type)
+                }
+            }
+            .pickerStyle(SegmentedPickerStyle())
+        }
+    }
+}
